refactor(stt): move pcm-math to shared for main+renderer reuse
